Enabling P2 and FC pilots for Channel estimation.
The P2 and FC Symbols are activated using the Frame builder and L1 signaling.
The FC pilots are extracted along with P2 pilots for any number of frames and are used for channel estimation and equalization.

The System parameters are changed to the following

FFT size 8K
Digital modulation 16 QAM
Scattered pilots PP1
Guard Interval 1/4

In this case we can have FC symbols as well as P2 symbols.