Edge Pilot Estimation and Channel Tracking
The Edge pilots were not used previously for channel estimation in the OMP algorithm. This has been changed and now edge pilots are also considered for channel estimation of the normal symbols

A channel tracking framework has been added. The normal symbols are equalized with the P2 symbols and the scattered pilots and edge pilots are used for channel tracking.
The channel tracking implemented so far is only amplitude tracking. A LS estimator is used for estimating the channel at the pilot positions and a linear interpolation is used to estimate the channel at the intermediate positions. After determining the complete channel it is equalized with a LS equalizer.

INPROGRESS: Adding phase tracking