Channel Tracking Updated
channel tracking is updated and now out of all the estimations done on the P2 preambles the most sparsest estimation is used for equalization before channel tracking.

The P2 data usage switch is changed from NoP2Data to P2Data so that it is more intuitive and easy to understand.